django-dashboard/requirements.txt
# Django Trading Dashboard Requirements
# Professional ML Trading Information Display Platform
# by Camlo Technologies

# Core Django
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Cache and Sessions
django-redis==5.4.0
redis==5.0.1

# Async and WebSockets
channels==4.0.0
channels-redis==4.1.0

# Forms and UI
django-crispy-forms==2.1
crispy-bootstrap5==0.7

# Environment and Configuration
python-decouple==3.8

# Utilities
django-extensions==3.2.3
django-filter==23.3

# CORS
django-cors-headers==4.3.1

# Static Files
whitenoise==6.6.0

# Background Tasks
celery==5.3.4
django-celery-beat==2.5.0

# Monitoring and Logging
sentry-sdk==1.38.0

# Data Processing
pandas==2.1.3
numpy==1.25.2

# ML and Analytics
scikit-learn==1.3.2
joblib==1.3.2

# HTTP Requests
requests==2.31.0

# Time and Date
python-dateutil==2.8.2

# Development Tools
django-debug-toolbar==4.2.0

# Production Server
gunicorn==21.2.0